# bugfix: timer gate, when updating the next wake up time, can suffer f…