…rom race condition, causing the imminent timer fired event to be lost.

performance impact:

after the change:
```
cadence|timer-bug ⇒ go test -cpu=1 -benchtime=10s -benchmem -run=^$ github.com/uber/cadence/service/history -bench ^BenchmarkLocalTimer$
goos: darwin
goarch: amd64
pkg: github.com/uber/cadence/service/history
BenchmarkLocalTimer 	100000000	       107 ns/op	       0 B/op	       0 allocs/op
PASS
ok  	github.com/uber/cadence/service/history	10.931s
cadence|timer-bug ⇒ go test -cpu=1 -benchtime=10s -benchmem -run=^$ github.com/uber/cadence/service/history -bench ^BenchmarkLocalTimer$
goos: darwin
goarch: amd64
pkg: github.com/uber/cadence/service/history
BenchmarkLocalTimer 	100000000	       100 ns/op	       0 B/op	       0 allocs/op
PASS
ok  	github.com/uber/cadence/service/history	10.205s
cadence|timer-bug ⇒ go test -cpu=1 -benchtime=10s -benchmem -run=^$ github.com/uber/cadence/service/history -bench ^BenchmarkLocalTimer$
goos: darwin
goarch: amd64
pkg: github.com/uber/cadence/service/history
BenchmarkLocalTimer 	100000000	       108 ns/op	       0 B/op	       0 allocs/op
PASS
ok  	github.com/uber/cadence/service/history	10.945s
cadence|timer-bug ⇒ go test -cpu=4 -benchtime=10s -benchmem -run=^$ github.com/uber/cadence/service/history -bench ^BenchmarkLocalTimer$
goos: darwin
goarch: amd64
pkg: github.com/uber/cadence/service/history
BenchmarkLocalTimer-4   	100000000	       211 ns/op	       0 B/op	       0 allocs/op
PASS
ok  	github.com/uber/cadence/service/history	21.347s
cadence|timer-bug ⇒ go test -cpu=4 -benchtime=10s -benchmem -run=^$ github.com/uber/cadence/service/history -bench ^BenchmarkLocalTimer$
goos: darwin
goarch: amd64
pkg: github.com/uber/cadence/service/history
BenchmarkLocalTimer-4   	100000000	       193 ns/op	       0 B/op	       0 allocs/op
PASS
ok  	github.com/uber/cadence/service/history	19.484s
cadence|timer-bug ⇒ go test -cpu=4 -benchtime=10s -benchmem -run=^$ github.com/uber/cadence/service/history -bench ^BenchmarkLocalTimer$
goos: darwin
goarch: amd64
pkg: github.com/uber/cadence/service/history
BenchmarkLocalTimer-4   	100000000	       236 ns/op	       0 B/op	       0 allocs/op
PASS
ok  	github.com/uber/cadence/service/history	23.771s
```

before the change:
```
cadence|master ⇒ go test -cpu=1 -benchtime=10s -benchmem -run=^$ github.com/uber/cadence/service/history -bench ^BenchmarkLocalTimer$
goos: darwin
goarch: amd64
pkg: github.com/uber/cadence/service/history
BenchmarkLocalTimer 	200000000	        88.8 ns/op	       0 B/op	       0 allocs/op
PASS
ok  	github.com/uber/cadence/service/history	26.831s
cadence|master⚡ ⇒ go test -cpu=1 -benchtime=10s -benchmem -run=^$ github.com/uber/cadence/service/history -bench ^BenchmarkLocalTimer$
goos: darwin
goarch: amd64
pkg: github.com/uber/cadence/service/history
BenchmarkLocalTimer 	200000000	        86.9 ns/op	       0 B/op	       0 allocs/op
PASS
ok  	github.com/uber/cadence/service/history	26.456s
cadence|master⚡ ⇒ go test -cpu=1 -benchtime=10s -benchmem -run=^$ github.com/uber/cadence/service/history -bench ^BenchmarkLocalTimer$
goos: darwin
goarch: amd64
pkg: github.com/uber/cadence/service/history
BenchmarkLocalTimer 	200000000	        88.1 ns/op	       0 B/op	       0 allocs/op
PASS
ok  	github.com/uber/cadence/service/history	26.589s
cadence|master⚡ ⇒ go test -cpu=4 -benchtime=10s -benchmem -run=^$ github.com/uber/cadence/service/history -bench ^BenchmarkLocalTimer$
goos: darwin
goarch: amd64
pkg: github.com/uber/cadence/service/history
BenchmarkLocalTimer-4   	100000000	       183 ns/op	       0 B/op	       0 allocs/op
PASS
ok  	github.com/uber/cadence/service/history	18.578s
cadence|master⚡ ⇒ go test -cpu=4 -benchtime=10s -benchmem -run=^$ github.com/uber/cadence/service/history -bench ^BenchmarkLocalTimer$
goos: darwin
goarch: amd64
pkg: github.com/uber/cadence/service/history
BenchmarkLocalTimer-4   	100000000	       185 ns/op	       0 B/op	       0 allocs/op
PASS
ok  	github.com/uber/cadence/service/history	18.653s
cadence|master⚡ ⇒ go test -cpu=4 -benchtime=10s -benchmem -run=^$ github.com/uber/cadence/service/history -bench ^BenchmarkLocalTimer$
goos: darwin
goarch: amd64
pkg: github.com/uber/cadence/service/history
BenchmarkLocalTimer-4   	100000000	       179 ns/op	       0 B/op	       0 allocs/op
PASS
ok  	github.com/uber/cadence/service/history	18.042s
```

the performance impact is ~<20% latency increase, which is negligible